Create README.md
Important Warning
This program is a remote administration tool (RAT). This program is created solely for scientific and educational purposes. The author is not responsible for any consequences associated with the use of this program.
Use of this program may violate laws and regulations in your jurisdiction. You are solely responsible for all activities performed using this program. The author of the program is not responsible for any loss, damage or negative consequences arising from the use or misuse of this program.